src/requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.0.1
django-cors-headers==2.1.0
django-filter==2.1.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
idna==2.8
Jinja2==2.10.1
ldap3==2.4.1
lxml==4.3.3
MarkupSafe==1.1.1
passlib==1.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
PyJWT==1.7.1
PyMySQL==0.8.0
python-docx==0.8.6
python-ldap==3.2.0
pytz==2019.1
raven==6.10.0
requests==2.21.0
SQLAlchemy==1.3.3
sqlparse==0.2.4
urllib3==1.24.2
